moving plot generator to general latex file
